requirements.txt
argcomplete==2.0.0
cachelib==0.10.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.0
Flask==2.2.2
Flask-Session==0.4.0
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.21.0
SQLAlchemy==1.4.46
Werkzeug==2.2.2
